OMS7-N: Allow Play Store to filter out OMS-only themes
This will be used in the separation of OMS and Legacy themes, and gets
added inside /system/etc/permissions.

Themers: To utilize this to create a theme only viewable on an OMS ROM,
add:

    <uses-feature
        android:name="projekt.substratum.theme"
        android:required="true" />

to the AndroidManifest.xml in the theme's APK. Play Store will limit the
theme to be viewed on OMS-only devices.

Change-Id: Ibcda2fb3ecd4004d2a3c12560b4e7bdc07505703
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>